requirements.txt
Django==4.2
djangorestframework==3.20.0
djangorestframework-authtoken==1.3.1
python-dotenv==1.0.0
